Algorithmic_Strategies/Crypto_Macro_Fundamental/requirements.txt
# Requirements for Crypto Macro-Fundamental Strategy

# Core packages
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0

# Hyperparameter Optimization
optuna>=3.3.0
optuna-dashboard>=0.13.0

# Data acquisition
yfinance>=0.2.28
pandas-datareader>=0.10.0
requests>=2.31.0

# CoinMarketCap API (unofficial)
python-coinmarketcap>=0.5
# Alternative: requests for custom API calls

# FRED API (Federal Reserve Economic Data)
fredapi>=0.5.1

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Visualization
plotly>=5.17.0

# Model interpretability
shap>=0.42.0

# Statistical analysis
statsmodels>=0.14.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# API keys configuration
# Create a .env file with:
# COINMARKETCAP_API_KEY=your_key_here
# FRED_API_KEY=your_key_here
